ci: give all artifacts individiual names
upload-artifacts v4 requires each upload in the workflow to have an individual name: https://github.com/actions/upload-artifact?tab=readme-ov-file#breaking-changes

The suffix for the related job is hence added, which was the case for the msvc test results already, and matches the XML file suffix as well. The test workflow picks up the artifacts with any suffix already.

Signed-off-by: MichaIng <micha@dietpi.com>